refactor(#3479): shrink idle_recap.rs below giant threshold
Extract two cohesive, behavior-preserving clusters from
src/services/discord/idle_recap.rs (1320 -> ~986 prod LoC, below the
1000-LoC giant threshold) into directory submodules:

- idle_recap/scrollback.rs (198 LoC): tmux capture-pane tail capture,
  the claude-e transcript-tail fallback (capture_transcript_scrollback +
  the unit-testable extract_transcript_tail_text / parse_transcript_line_text
  workers), and the Haiku summarize_with_haiku call.
- idle_recap/context_display.rs (154 LoC): the live/latest-turn/session
  token selection state machine (select_recap_context + RecapContextDisplay),
  freshness/provider-match guards, and the token/duration formatters.

Bodies moved verbatim. Submodules use `use super::*;`. Symbols the parent
prod code or in-file tests reach are pub(super); the three scrollback entry
points stay pub(crate) and are re-exported so server::routes::idle_recap
keeps byte-identical idle_recap::<fn> call sites. The parent's in-file test
module is unchanged and reaches the moved items unqualified via `use super::*`.

Graduate the file from the giant-file gate: removed its [[entry]] from
giant_file_registry.toml and its ratchet baseline from
audit_maintainability_giant_baseline.toml, regenerated ARCHITECTURE.md, and
updated change-surfaces.md.
