+++ scripts/giant_file_registry.toml
@@ -303,17 +303,11 @@ owner = "discord-finalizer"
 deadline = "2026-08-31"
 decompose_issue = "#3016"
 
-[[entry]]
-# Crossed the 1000-prod-LoC threshold when #3146 Part 1 added the idle-recap
-# active-turn gating: the shared clear helpers (capture-at-claim + CAS), the
-# `channel_has_active_turn` mailbox/inflight probe, and the `post_recheck_action`
-# seam. The card-lifecycle helpers (compose/post/clear/CAS/probe) are cohesive
-# but sizable; decompose the recap module into compose vs lifecycle/clear
-# submodules under #3036.
-file = "src/services/discord/idle_recap.rs"
-owner = "discord-recap"
-deadline = "2026-08-31"
-decompose_issue = "#3405"
+# NOTE: src/services/discord/idle_recap.rs previously had an [[entry]] here; #3479
+# extracted the scrollback-capture/Haiku-summarizer and token-context-display
+# clusters into `idle_recap/scrollback.rs` and `idle_recap/context_display.rs`,
+# dropping its production LoC below the giant threshold. Its entry (and ratchet
+# baseline) were removed to avoid a ghost registration.
 
 # NOTE: src/services/discord/turn_bridge/terminal_delivery.rs previously had an
 # [[entry]] here; the #3028 splitter fix dropped its production LoC below the
