+++ src/services/discord/idle_recap.rs
@@ -26,6 +26,25 @@ use tokio::task;
 
 use crate::services::provider::{CancelToken, ProviderKind};
 
+mod context_display;
+mod scrollback;
+
+// Scrollback capture / Haiku summarizer (#3479): re-exported so external
+// callers (`server::routes::idle_recap`) keep the `idle_recap::<fn>` path,
+// and the in-file test module reaches them unqualified via `use super::*`.
+pub(crate) use self::scrollback::{
+    capture_tmux_scrollback, capture_transcript_scrollback, summarize_with_haiku,
+};
+#[cfg(test)]
+use self::scrollback::{extract_transcript_tail_text, parse_transcript_line_text};
+
+// Context-display helpers (#3479) used by `compose_recap_header` /
+// `attach_live_context_usage`; the test module reaches the rest unqualified.
+use self::context_display::{
+    RecapContextDisplay, format_korean_duration, format_token_count, normalized_text,
+    provider_session_ids, select_recap_context,
+};
+
 const FALLBACK_CONTEXT_WINDOW_TOKENS: u64 = 128_000;
 const SESSION_TOKEN_FRESHNESS_MAX_SECS: i64 = 30 * 60;
 
@@ -336,339 +355,6 @@ fn compose_recap_header(snapshot: &RecapSnapshot) -> String {
         }
         RecapContextDisplay::Stale => format!("📦 context stale · idle {idle_since}"),
         RecapContextDisplay::Unknown => format!("📦 context unknown · idle {idle_since}"),
-    }
-}
-
-/// Best-effort tail capture of the live tmux pane via `tmux capture-pane`.
-/// Returns `None` if the session is gone or the binary is unavailable —
-/// the caller treats that as "no scrollback, post header-only".
-pub(crate) async fn capture_tmux_scrollback(session_name: &str) -> Option<String> {
-    let session = session_name.to_string();
-    task::spawn_blocking(move || {
-        std::process::Command::new("tmux")
-            .args([
-                "capture-pane",
-                "-p",
-                "-J",
-                "-S",
-                &format!("-{TMUX_SCROLLBACK_LINES}"),
-                "-t",
-                &session,
-            ])
-            .output()
-            .ok()
-            .and_then(|out| {
-                if out.status.success() {
-                    Some(String::from_utf8_lossy(&out.stdout).trim().to_string())
-                } else {
-                    None
-                }
-            })
-    })
-    .await
-    .ok()
-    .flatten()
-    .filter(|s| !s.is_empty())
-}
-
-/// Fallback scrollback source for runtimes without a live tmux pane —
-/// notably the `claude-e` per-turn spawn runtime, which never attaches a
-/// long-lived tmux session. Reads the Claude transcript JSONL at
-/// `~/.claude/projects/<encoded-cwd>/<session_id>.jsonl`, parses each line,
-/// and emits the last ~`TMUX_SCROLLBACK_LINES` user/assistant text turns in
-/// a `[role] text` shape that the recap summarizer can consume the same
-/// way it consumes tmux scrollback.
-///
-/// Returns `None` when the transcript is missing, unreadable, contains no
-/// human-readable turns, or `session_id` is not a valid UUID. The recap
-/// pipeline degrades gracefully to a header-only card in that case.
-///
-/// As a free bonus this also covers stale tmux sessions whose pane has
-/// already been torn down: the transcript file outlives the tmux pane.
-pub(crate) async fn capture_transcript_scrollback(
-    cwd: &std::path::Path,
-    session_id: &str,
-) -> Option<String> {
-    let transcript_path =
-        crate::services::claude_tui::transcript_tail::claude_transcript_path(cwd, session_id, None)
-            .ok()?;
-    let path_for_blocking = transcript_path.clone();
-    task::spawn_blocking(move || extract_transcript_tail_text(&path_for_blocking))
-        .await
-        .ok()
-        .flatten()
-}
-
-/// Synchronous worker for `capture_transcript_scrollback`. Splits out so
-/// the parsing logic is unit-testable without an async runtime.
-fn extract_transcript_tail_text(transcript_path: &std::path::Path) -> Option<String> {
-    use std::collections::VecDeque;
-    use std::io::BufRead;
-
-    let file = std::fs::File::open(transcript_path).ok()?;
-    let reader = std::io::BufReader::new(file);
-    let cap = TMUX_SCROLLBACK_LINES as usize;
-    let mut buf: VecDeque<String> = VecDeque::with_capacity(cap);
-    for line in reader.lines().map_while(Result::ok) {
-        let Some(entry) = parse_transcript_line_text(&line) else {
-            continue;
-        };
-        if buf.len() == cap {
-            buf.pop_front();
-        }
-        buf.push_back(entry);
-    }
-    if buf.is_empty() {
-        None
-    } else {
-        Some(buf.into_iter().collect::<Vec<_>>().join("\n"))
-    }
-}
-
-/// Extract a `[role] text` line from a single Claude transcript JSONL row.
-/// Returns `None` for rows without human-readable content (init/done/status,
-/// tool uses, tool results, attachments, etc.) so the recap summarizer
-/// only sees signal.
-fn parse_transcript_line_text(line: &str) -> Option<String> {
-    let trimmed = line.trim();
-    if trimmed.is_empty() {
-        return None;
-    }
-    let value: serde_json::Value = serde_json::from_str(trimmed).ok()?;
-    let role = match value.get("type")?.as_str()? {
-        "user" => "user",
-        "assistant" => "assistant",
-        _ => return None,
-    };
-    let content = value.get("message")?.get("content")?;
-    let text = match content {
-        serde_json::Value::String(s) => s.trim().to_string(),
-        serde_json::Value::Array(blocks) => {
-            let mut parts = Vec::new();
-            for block in blocks {
-                if block.get("type").and_then(|t| t.as_str()) != Some("text") {
-                    continue;
-                }
-                if let Some(piece) = block.get("text").and_then(|t| t.as_str()) {
-                    let piece = piece.trim();
-                    if !piece.is_empty() {
-                        parts.push(piece.to_string());
-                    }
-                }
-            }
-            parts.join(" ")
-        }
-        _ => return None,
-    };
-    let text = text.trim();
-    if text.is_empty() {
-        return None;
-    }
-    Some(format!("[{role}] {text}"))
-}
-
-/// Ask Claude Haiku for a 1-2 sentence Korean recap. Time-bounded; returns
-/// `None` on any failure so the caller can fall back to a header-only card.
-///
-/// Previously this routed to a local `opencode serve` (Gemma 27B) build,
-/// but resident memory on the mac-book host was the bottleneck. Haiku 4.5
-/// is cheap enough per call (a few cents per million tokens) and fast
-/// enough on remote API that it comfortably fits inside the 20s budget
-/// without holding any RAM on the host.
-///
-/// The Claude call is wrapped in `spawn_blocking`. A `tokio::time::timeout`
-/// alone would only cancel the *await* on the JoinHandle and leave the
-/// blocking thread running with a live `claude` subprocess. So we also
-/// pass a `CancelToken` into the Claude wrapper and *signal it* from the
-/// timeout watchdog. The Claude simple-cancel watcher polls
-/// `cancel_requested` and tears down the child process tree as soon as it
-/// sees the flag.
-pub(crate) async fn summarize_with_haiku(scrollback: &str) -> Option<String> {
-    if scrollback.is_empty() {
-        return None;
-    }
-    let prompt = format!(
-        "다음은 AI 코딩 에이전트와 사용자의 마지막 대화 ~100줄입니다. \
-         사용자가 지금 다시 돌아왔을 때 \"어떤 작업을 하던 중이었는지\"를 \
-         즉시 기억할 수 있도록 1-2문장으로 한국어 요약을 만드세요. \
-         도구 호출 / 스크롤 / 진행 표시 같은 노이즈는 무시하고 \
-         실제 작업 내용(파일·결정·다음 단계)에 집중하세요. \
-         결과만 출력하고 다른 말은 붙이지 마세요.\n\n---\n\n{scrollback}",
-    );
-
-    let cancel = std::sync::Arc::new(CancelToken::new());
-    let cancel_for_blocking = cancel.clone();
-    let join = task::spawn_blocking(move || {
-        crate::services::claude::execute_command_simple_cancellable_with_model(
-            &prompt,
-            Some(RECAP_SUMMARY_MODEL),
-            Some(cancel_for_blocking),
-        )
-    });
-
-    let result = match tokio::time::timeout(RECAP_SUMMARY_TIMEOUT, join).await {
-        Ok(join_result) => match join_result {
-            Ok(Ok(text)) => text,
-            Ok(Err(_)) => return None,
-            Err(_) => return None,
-        },
-        Err(_) => {
-            // Timeout fired. Signal the cancel token so the blocking
-            // closure exits at the next Claude wrapper poll and the
-            // spawned child tree is reaped.
-            cancel.cancel_with_tmux_cleanup();
-            return None;
-        }
-    };
-
-    let trimmed = result.trim();
-    if trimmed.is_empty() {
-        None
-    } else {
-        Some(trimmed.to_string())
-    }
-}
-
-fn format_token_count(n: u64) -> String {
-    if n >= 1_000_000 {
-        format!("{:.1}M", n as f64 / 1_000_000.0)
-    } else if n >= 1_000 {
-        format!("{:.1}k", n as f64 / 1_000.0)
-    } else {
-        n.to_string()
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum RecapContextDisplay {
-    Known { used: u64, window: u64 },
-    Stale,
-    Unknown,
-}
-
-fn select_recap_context(snapshot: &RecapSnapshot, now: DateTime<Utc>) -> RecapContextDisplay {
-    let known = |used, window| RecapContextDisplay::Known {
-        used: display_context_tokens(snapshot, used, window),
-        window,
-    };
-    if let Some(live) = snapshot
-        .live_context_usage
-        .filter(|live| live.used_tokens > 0 && live.context_window_tokens > 0)
-    {
-        return known(live.used_tokens, live.context_window_tokens);
-    }
-
-    let window = context_window_for(snapshot);
-    if let Some(used) = latest_turn_context_tokens(snapshot) {
-        return known(used, window);
-    }
-    if session_tokens_are_stale_or_incompatible(snapshot, now) {
-        return RecapContextDisplay::Stale;
-    }
-    fresh_session_tokens(snapshot, now)
-        .map_or(RecapContextDisplay::Unknown, |used| known(used, window))
-}
-
-fn display_context_tokens(snapshot: &RecapSnapshot, used: u64, window: u64) -> u64 {
-    match ProviderKind::from_str(&snapshot.provider) {
-        Some(ProviderKind::Codex) if window > 0 => used.min(window),
-        _ => used,
-    }
-}
-
-fn latest_turn_context_tokens(snapshot: &RecapSnapshot) -> Option<u64> {
-    latest_turn_matches_active_session(snapshot).then_some(())?;
-    let input = non_negative_i64_to_u64(snapshot.latest_turn_input_tokens?)?;
-    let used = input
-        .saturating_add(non_negative_i64_to_u64(
-            snapshot.latest_turn_cache_create_tokens.unwrap_or(0),
-        )?)
-        .saturating_add(non_negative_i64_to_u64(
-            snapshot.latest_turn_cache_read_tokens.unwrap_or(0),
-        )?);
-    (used > 0).then_some(used)
-}
-
-fn latest_turn_matches_active_session(snapshot: &RecapSnapshot) -> bool {
-    snapshot.latest_turn_finished_at.is_some()
-        && same_normalized_opt(
-            snapshot.latest_turn_provider.as_deref(),
-            Some(snapshot.provider.as_str()),
-        )
-        && (same_normalized_opt(
-            snapshot.latest_turn_session_key.as_deref(),
-            Some(snapshot.session_key.as_str()),
-        ) || snapshot
-            .latest_turn_session_id
-            .as_deref()
-            .and_then(normalized_text)
-            .is_some_and(|latest| provider_session_ids(snapshot).any(|active| active == latest)))
-}
-
-fn provider_session_ids(snapshot: &RecapSnapshot) -> impl Iterator<Item = &str> {
-    [
-        snapshot.claude_session_id.as_deref(),
-        snapshot.raw_provider_session_id.as_deref(),
-    ]
-    .into_iter()
-    .flatten()
-    .filter_map(normalized_text)
-}
-
-fn session_tokens_are_stale_or_incompatible(snapshot: &RecapSnapshot, now: DateTime<Utc>) -> bool {
-    let Some(tokens) = snapshot.tokens.filter(|value| *value > 0) else {
-        return false;
-    };
-    non_negative_i64_to_u64(tokens).is_none()
-        || (snapshot.latest_turn_finished_at.is_some()
-            && !latest_turn_matches_active_session(snapshot))
-        || !session_tokens_are_fresh(snapshot, now)
-}
-
-fn fresh_session_tokens(snapshot: &RecapSnapshot, now: DateTime<Utc>) -> Option<u64> {
-    let tokens = non_negative_i64_to_u64(snapshot.tokens?)?;
-    (tokens > 0 && session_tokens_are_fresh(snapshot, now)).then_some(tokens)
-}
-
-fn session_tokens_are_fresh(snapshot: &RecapSnapshot, now: DateTime<Utc>) -> bool {
-    snapshot.tokens_updated_at.is_some_and(|updated_at| {
-        let age = now - updated_at;
-        age.num_seconds() >= 0 && age.num_seconds() <= SESSION_TOKEN_FRESHNESS_MAX_SECS
-    })
-}
-
-fn non_negative_i64_to_u64(value: i64) -> Option<u64> {
-    u64::try_from(value).ok()
-}
-
-fn same_normalized_opt(left: Option<&str>, right: Option<&str>) -> bool {
-    matches!(
-        (left.and_then(normalized_text), right.and_then(normalized_text),),
-        (Some(left), Some(right)) if left.eq_ignore_ascii_case(right)
-    )
-}
-
-fn normalized_text(value: &str) -> Option<&str> {
-    let trimmed = value.trim();
-    (!trimmed.is_empty()).then_some(trimmed)
-}
-
-fn context_window_for(snapshot: &RecapSnapshot) -> u64 {
-    ProviderKind::from_str(&snapshot.provider).map_or(FALLBACK_CONTEXT_WINDOW_TOKENS, |provider| {
-        provider.resolve_context_window(snapshot.model.as_deref())
-    })
-}
-
-fn format_korean_duration(dur: chrono::Duration) -> String {
-    let secs = dur.num_seconds().max(0);
-    if secs >= 86_400 {
-        format!("{}일", secs / 86_400)
-    } else if secs >= 3_600 {
-        format!("{}시간", secs / 3_600)
-    } else if secs >= 60 {
-        format!("{}분", secs / 60)
-    } else {
-        format!("{}초", secs)
     }
 }
